Provide extra build instructions and fix build failures
An attempt to build on Ubuntu 20.04 using the command
`./configure ; make` failed because of 2 main reasons:
1. `autoconf` had been run with a previous version of
autoconf that generated build scripts that expected a
specific version of aclocal. The build was therefore
failing with the error:

    `line 81: aclocal-1.11: command not found`

2. Once the first error was resolved by reruning
autoconf with the command `autoreconf -fis -I m4`,
it was discovered that now the build was failing
because one of the tests was named `array.cc`.
Libtool was generating a bash script called
`array` which the build system was attempting
to include in the STL tuple header file during
preprocessing. Since this file was a bash file
and not the expected STL header file array, the
build was failing with the below error:

```
In file included from /usr/include/c++/9/tuple:39,
                 from /usr/include/c++/9/bits/stl_map.h:63,
                 from /usr/include/c++/9/map:61,
                 from ../../include/TaskIRSuif.h:6,
                 from ../../include/TaskIR.h:2,
                 from ../../include/TaskGraph:5,
                 from loops.cc:2:
./array:1:2: error: invalid preprocessing directive #!
    1 | #! /bin/bash

```

To fix issue 1 above, the build artifacts are removed
from the git repo and the README.md is updated to instruct
the user to first run autoreconf and automake  before
running `./configure ; make`. autoreconf and automake
set up the users build scripts to reference tools
installed in the users system, rather than hard coded
ones that were initially included in the repository.

To fix issue 2, we rename the array test to
array_test so that it does not cause the libtool
generated bash script named array to clash with
the C++ STL header array.